@@ -1,7 +1,7 @@
 {
   "name": "monaco-editor",
   "private": true,
-  "version": "0.15.6",
+  "version": "0.22.3",
   "description": "A browser based code editor",
   "author": "Microsoft Corporation",
   "license": "MIT",
@@ -18,21 +18,21 @@
     "url": "https://github.com/Microsoft/monaco-editor"
   },
   "devDependencies": {
-    "clean-css": "^4.2.1",
-    "event-stream": "3.3.4",
-    "gulp": "^3.9.1",
-    "gulp-typedoc": "^2.2.0",
-    "http-server": "^0.11.1",
-    "monaco-css": "2.3.0",
-    "monaco-editor-core": "0.15.5",
-    "monaco-html": "2.3.0",
-    "monaco-json": "2.3.0",
-    "monaco-languages": "1.6.0",
-    "monaco-typescript": "3.3.1",
-    "rimraf": "^2.6.2",
-    "typedoc": "^0.11.1",
-    "typescript": "^3.0.1",
-    "uncss": "^0.16.2",
-    "vinyl": "^0.5.3"
+    "clean-css": "^4.2.3",
+    "event-stream": "4.0.1",
+    "gulp": "^4.0.2",
+    "gulp-typedoc": "^2.2.5",
+    "monaco-css": "3.2.0",
+    "monaco-editor-core": "0.22.3",
+    "monaco-html": "3.2.0",
+    "monaco-json": "3.3.0",
+    "monaco-languages": "2.2.0",
+    "monaco-typescript": "4.2.0",
+    "rimraf": "^3.0.2",
+    "typedoc": "^0.19.1",
+    "typescript": "4.1.3",
+    "uncss": "^0.17.3",
+    "vinyl": "^2.2.1",
+    "yaserver": "^0.3.0"
   }
 }
