@@ -1,4 +1,4 @@
-function createDependencyProposals() {
+function createDependencyProposals(range) {
     // returning a static list of proposals, not even looking at the prefix (filtering is done by the Monaco editor),
     // here you could do a server side lookup
     return [
@@ -6,19 +6,30 @@ function createDependencyProposals() {
             label: '"lodash"',
             kind: monaco.languages.CompletionItemKind.Function,
             documentation: "The Lodash library exported as Node.js modules.",
-            insertText: '"lodash": "*"'
+            insertText: '"lodash": "*"',
+            range: range
         },
         {
             label: '"express"',
             kind: monaco.languages.CompletionItemKind.Function,
             documentation: "Fast, unopinionated, minimalist web framework",
-            insertText: '"express": "*"'
+            insertText: '"express": "*"',
+            range: range
         },
         {
             label: '"mkdirp"',
             kind: monaco.languages.CompletionItemKind.Function,
             documentation: "Recursively mkdir, like <code>mkdir -p</code>",
-            insertText: '"mkdirp": "*"'
+            insertText: '"mkdirp": "*"',
+            range: range
+        },
+        {
+            label: '"my-third-party-library"',
+            kind: monaco.languages.CompletionItemKind.Function,
+            documentation: "Describe your library here",
+            insertText: '"${1:my-third-party-library}": "${2:1.2.3}"',
+            insertTextRules: monaco.languages.CompletionItemInsertTextRule.InsertAsSnippet,
+            range: range
         }
     ];
 }
@@ -28,10 +39,19 @@ monaco.languages.registerCompletionItemProvider('json', {
     provideCompletionItems: function(model, position) {
         // find out if we are completing a property in the 'dependencies' object.
         var textUntilPosition = model.getValueInRange({startLineNumber: 1, startColumn: 1, endLineNumber: position.lineNumber, endColumn: position.column});
-        var match = textUntilPosition.match(/"dependencies"\s*:\s*{\s*("[^"]*"\s*:\s*"[^"]*"\s*,\s*)*("[^"]*)?$/);
-        var suggestions = match ? createDependencyProposals() : [];
+        var match = textUntilPosition.match(/"dependencies"\s*:\s*\{\s*("[^"]*"\s*:\s*"[^"]*"\s*,\s*)*([^"]*)?$/);
+        if (!match) {
+            return { suggestions: [] };
+        }
+        var word = model.getWordUntilPosition(position);
+        var range = {
+            startLineNumber: position.lineNumber,
+            endLineNumber: position.lineNumber,
+            startColumn: word.startColumn,
+            endColumn: word.endColumn
+        };
         return {
-            suggestions: suggestions
+            suggestions: createDependencyProposals(range)
         };
     }
 });
