@@ -4,10 +4,11 @@
 
 	'use strict';
 
+	var isMac = /Mac/i.test(navigator.userAgent);
 	window.onload = function () {
 		require(['vs/editor/editor.main'], function () {
 			xhr('playground/monaco.d.ts.txt').then(function (response) {
-				monaco.languages.typescript.javascriptDefaults.addExtraLib(response.responseText, 'monaco.d.ts');
+				monaco.languages.typescript.javascriptDefaults.addExtraLib(response.responseText, 'ts:monaco.d.ts');
 				monaco.languages.typescript.javascriptDefaults.addExtraLib([
 					'declare var require: {',
 					'	toUrl(path: string): string;',
@@ -16,7 +17,7 @@
 					'	config(data: any): any;',
 					'	onError: Function;',
 					'};',
-				].join('\n'), 'require.d.ts');
+				].join('\n'), 'ts:require.d.ts');
 			});
 
 			var loading = document.getElementById('loading');
@@ -157,8 +158,11 @@
 			htmlTab.onclick = function () { changeTab(htmlTab, 'html'); };
 			tabArea.appendChild(htmlTab);
 
-			var runBtn = document.createElement('span');
+			var runLabel = 'Press ' + (isMac ? 'CMD + return' : 'CTRL + Enter') + ' to run the code.';
+			var runBtn = document.createElement('button');
 			runBtn.className = 'action run';
+			runBtn.setAttribute('role', 'button');
+			runBtn.setAttribute('aria-label', runLabel);
 			runBtn.appendChild(document.createTextNode('Run'));
 			runBtn.onclick = function () { run(); };
 			tabArea.appendChild(runBtn);
@@ -307,6 +311,20 @@
 		function run() {
 			doRun(runContainer);
 		}
+
+		editor.addCommand(monaco.KeyMod.CtrlCmd | monaco.KeyCode.Enter, run);
+		window.addEventListener('keydown', function keyDown(ev) {
+			if ((isMac && !ev.metaKey) || !ev.ctrlKey) {
+				return;
+			}
+
+			if (ev.shiftKey || ev.altKey || ev.keyCode !== 13) {
+				return;
+			}
+
+			ev.preventDefault();
+			run();
+		});
 	}
 
 	var runIframe = null, runIframeHeight = 0;
